+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Couple of small updates in Tooltip and Modal.
+
 ## [0.26.1] - 2020-10-05
 
 - Hotfix for tooltip in withCopyToClipboard HOC.